% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_files.R
\name{get_nhanes_filenames}
\alias{get_nhanes_filenames}
\title{Get NHANES filenames}
\usage{
get_nhanes_filenames(setup, save_file_list = TRUE)
}
\arguments{
\item{setup}{The object (list) returned from the "setup_nhanes" function}

\item{save_file_list}{Indicates whether the directory contents should be saved as a separate file (called "download_file_specs.rds")}
}
\value{
A character vector of ftp urls ("filenames") to download
}
\description{
Gets all relevant NHANES filenames for a given year, both for the main data and for the mortality files
}
\examples{
# NOT RUN
# assumes setup_nhanes has been run and its result assigned to object "n"
# filenames <- get_nhanes_filenames(setup = n)
}

