% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{get_nhanes_listing}
\alias{get_nhanes_listing}
\title{Generate a list of all downloadable NHANES files}
\usage{
get_nhanes_listing()
}
\value{
A dataframe with a list of all of the available NHANES files.
}
\description{
Generates a list of downloadable NHANES data files from the CDC website, as well as the meta-data about the files.  This is primarily used to populate the internal table in the package.  However, this function is accessible so you can compare the most current version to the internal list within the package to see if anything has changed.  Note that changes to the NHANES website might make this function fail.
}

