% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{load_labs_merge}
\alias{load_labs_merge}
\title{Create data dictionary}
\usage{
load_labs_merge(vec_of_files = NULL, yr)
}
\arguments{
\item{vec_of_files}{The character vector of files to be retrieved.  The "demo" file is ALWAYS included, and should NOT be specified.}

\item{yr}{The year for which the file should be extracted.}
}
\value{
Returns a dataframe (a data.table) of all of the labels from each file in the character vector, including "demo".
}
\description{
Creates a simple data dictionary based on variable labels for each NHANES file
}

