% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{load_merge}
\alias{load_merge}
\title{Load and merge NHANES files}
\usage{
load_merge(vec_of_files, yr)
}
\arguments{
\item{vec_of_files}{A character vector of NHANES files (e.g., c("mcq", "biopro")) that identifies the stem of the desired file(s).  The demo file is ALWAYS included because it has the survey weights.  This vector should not include the final letter (e.g., _c) that indicates the wave (see yr).}

\item{yr}{The year for which the file should be extracted.}
}
\value{
Returns a dataframe (which is also a data.table) with one column for each variable in each file requested.
}
\description{
Takes vector of NHANES file names, loads them, and merges them all by SEQN.  Automatically loads demo, DO NOT include "demo" in the character vector.
}
\examples{
# Example:  load many files listed in character vector
# listing <- c("mcq", "dex", "hcq", "hiq", "vix", "uc") # demo is assumed
# full <- load_merge(listing, 2003)
# Example:  load label files listed in character vector
# full_labels <- load_labs_merge(listing, 2003)
}

