% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{nhanes_design}
\alias{nhanes_design}
\title{Generate NHANES design object}
\usage{
nhanes_design(df, wt = "WTINT2YR")
}
\arguments{
\item{df}{a data.frame or data.table created from NHANES data}

\item{wt}{The weight variable in the data as a character vector.  Common options include WTINT2YR, WTMEC2YR, and WTSA2YR.  See NHANES documentation for assigning weights across multiple years, and for choosing weights when there is more than 1 sampling frame (e.g., interview, mobile exam, fasting laboratory values, and other data subsets)}
}
\value{
Returns a survey design object (which contains all of the data) for use in analyses
}
\description{
Takes a data.frame with NHANES data and at least one weight variable, and creates the proper design object for it.  It simply puts the proper PSU strata, nest, and user-supplied weight into the svydesign function.  Note this does NOT set options(survey.lonely.psu = "adjust").
}
\examples{
\dontrun{
nhanes_design(df = df, wt = "WTINT2YR")
}
}

