% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{output_table}
\alias{output_table}
\title{Print mean, SE, and CI nicely from survey functions}
\usage{
output_table(t, dig = 0)
}
\arguments{
\item{t}{A survey object from a svytotal, svymean, or svyby function}

\item{dig}{The number of digits to the right of the decimal (using formatC)}
}
\value{
Returns a dataframe with character-formatted results summarizing the mean, standard error, and confidence interval
}
\description{
Takes a survey object and prints out the results in a nicer format for a report.
}
\examples{
\dontrun{
# nested syntax of base R:
out_tab(
   svytotal(~ variable_of_interest, survey_design_object, na.rm = TRUE)
)

# can be used with magrittr syntax as well
svytotal(~ variable_of_interest, survey_design_object, na.rm = TRUE) \%>\%
   out_tab
}
}

